+++ allennlp/training/__init__.py
@@ -1,6 +1,5 @@
 from allennlp.training.checkpointer import Checkpointer
 from allennlp.training.tensorboard_writer import TensorBoardWriter
-from allennlp.training.console_writer import ConsoleWriter
 from allennlp.training.log_writer import LogWriter
 from allennlp.training.no_op_trainer import NoOpTrainer
 from allennlp.training.trainer import (