+++ CHANGELOG.md
@@ -20,8 +20,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `ModelUsage` to `ModelCard` class.
 - Added a way to specify extra parameters to the predictor in an `allennlp predict` call.
 - Added a way to initialize a `Vocabulary` from transformers models.
-- Added `LogWriter` and `ConsoleWriter` classes. `TensorBoardWriter` and `ConsoleWriter` both inherit from `LogWriter`.
-- Added `LogCallback` and `ConsoleLogCallback` classes. `TensorBoardCallback` and `ConsoleLogCallback` both inherit from `LogCallback`.
+- Added `LogWriter` classe. `TensorBoardWriter` now inherits from `LogWriter`.
+- Added `LogCallback` and `ConsoleLogCallback` classes. `TensorBoardCallback` inherits from `LogCallback`.
 - Added an example for fields of type `ListField[TextField]` to `apply_token_indexers` API docs.
 
 ### Fixed